Develop

+++ app/Models/V1/MicrocontrollerData.php
@@ -151,26 +151,49 @@ class MicrocontrollerData extends Model
             $json['varCh_interval'] = 0;
             $json['varLh_interval'] = 0;
         } else {
-            $last_data = $client->microcontrollerData()->orderBy('source_timestamp', 'desc')->first();
-            if ($last_data) {
-                if (new Carbon($last_data->source_timestamp) >= $current_time  ){
-                    $this->delete();
-                    return;
-                }
-                $last_raw_json = json_decode($last_data->raw_json, true);
-                if ($json['import_wh'] <= 0) {
+            if ($flag) {
+                $last_data = $client->microcontrollerData()->orderBy('source_timestamp', 'desc')->first();
+                if ($last_data) {
+                    $last_raw_json = json_decode($last_data->raw_json, true);
+                    if ($json['import_wh'] <= 0) {
 
-                    if ($last_raw_json['import_wh']>0) {
-                        $this->forceDelete();
-                        return;
+                        if ($last_raw_json['import_wh'] > 0) {
+                            $this->forceDelete();
+                            return;
+                        }
                     }
-                }/*
-                if ($json['import_wh'] < $last_raw_json['import_wh']) {
-                    $json['import_wh'] = $last_raw_json['import_wh'];
                 }
-                if ($json['import_VArh'] < $last_raw_json['import_VArh']) {
-                    $json['import_VArh'] = $last_raw_json['import_VArh'];
-                }*/
+            } else{
+                $last_data = $client->microcontrollerData()->where('source_timestamp', '<', $current_time->format('Y-m-d H:i:s'))->orderBy('source_timestamp', 'desc')->first();
+                if ($last_data) {
+                    $last_raw_json = json_decode($last_data->raw_json, true);
+                    if ($json['import_wh'] <= 0) {
+
+                        if ($last_raw_json['import_wh'] > 0) {
+                            $this->forceDelete();
+                            return;
+                        }
+                    }
+                } else{
+                    $json['kwh_interval'] = 0;
+                    $json['varh_interval'] = 0;
+                    $json['varCh_acumm'] = floatval($json['ph1_varCh_acumm']) + floatval($json['ph2_varCh_acumm']) + floatval($json['ph3_varCh_acumm']);
+                    $json['varLh_acumm'] = floatval($json['ph1_varLh_acumm']) + floatval($json['ph2_varLh_acumm']) + floatval($json['ph3_varLh_acumm']);
+                    $json['ph1_varCh_interval'] = 0;
+                    $json['ph1_varLh_interval'] = 0;
+                    $json['ph2_varCh_interval'] = 0;
+                    $json['ph2_varLh_interval'] = 0;
+                    $json['ph3_varCh_interval'] = 0;
+                    $json['ph3_varLh_interval'] = 0;
+                    $json['ph1_kwh_interval'] = 0;
+                    $json['ph2_kwh_interval'] = 0;
+                    $json['ph3_kwh_interval'] = 0;
+                    $json['ph1_varh_interval'] = 0;
+                    $json['ph2_varh_interval'] = 0;
+                    $json['ph3_varh_interval'] = 0;
+                    $json['varCh_interval'] = 0;
+                    $json['varLh_interval'] = 0;
+                }
             }
             $reference_hour = $current_time->copy()->subHour();
             $reference_data = $client->microcontrollerData()
@@ -181,9 +204,8 @@ class MicrocontrollerData extends Model
                 $reference_data = $client->microcontrollerData()->where('source_timestamp', '<', $reference_hour->format('Y-m-d H:00:00'))->orderBy('source_timestamp', 'desc')->first();
             }
             if (empty($reference_data)) {
-
-
-                $json['kwh_interval'] = $json['import_wh'] - $last_raw_json['import_wh'];
+                if ($last_data) {
+                    $json['kwh_interval'] = $json['import_wh'] - $last_raw_json['import_wh'];
                     $json['varh_interval'] = $json['import_VArh'] - $last_raw_json['import_VArh'];
                     $json['varCh_acumm'] = floatval($json['ph1_varCh_acumm']) + floatval($json['ph2_varCh_acumm']) + floatval($json['ph3_varCh_acumm']);
                     $json['varLh_acumm'] = floatval($json['ph1_varLh_acumm']) + floatval($json['ph2_varLh_acumm']) + floatval($json['ph3_varLh_acumm']);
@@ -209,11 +231,10 @@ class MicrocontrollerData extends Model
                     $json['ph3_varh_interval'] = $json['ph3_import_kvarh'] - $last_raw_json['ph3_import_kvarh'];
                     $json['varCh_interval'] = $json['varCh_acumm'] - floatval($last_raw_json['varCh_acumm']);
                     $json['varLh_interval'] = $json['varLh_acumm'] - floatval($last_raw_json['varLh_acumm']);
+                }
 
             } else {
-                echo $this->id."\n";
-                echo $last_data->id."\n";
-
+                if ($last_data) {
                     $reference_data_json = json_decode($reference_data->raw_json, true);
                     $json['kwh_interval'] = $json['import_wh'] - $reference_data_json['import_wh'];
                     $json['varh_interval'] = $json['import_VArh'] - $reference_data_json['import_VArh'];
@@ -241,10 +262,9 @@ class MicrocontrollerData extends Model
                     $json['ph3_varh_interval'] = $json['ph3_import_kvarh'] - $reference_data_json['ph3_import_kvarh'];
                     $json['varCh_interval'] = $json['varCh_acumm'] - floatval($reference_data_json['varCh_acumm']);
                     $json['varLh_interval'] = $json['varLh_acumm'] - floatval($reference_data_json['varLh_acumm']);
-
+                }
             }
         }
-
         $this->client_id = $client->id;
         $this->accumulated_real_consumption = $json['import_wh'];
         $this->interval_real_consumption = $json['kwh_interval'];
@@ -257,17 +277,25 @@ class MicrocontrollerData extends Model
         $this->raw_json = $json;
         if (!$flag) {
             if ($client->microcontrollerData()->where('source_timestamp', $current_time->format('Y-m-d H:i:s'))->exists()) {
-                if ($this->hourlyMicrocontrollerData()->exists()) {
-                    $this->hourlyMicrocontrollerData()->forceDelete();
+                $equals = $client->microcontrollerData()->where('source_timestamp', $current_time->format('Y-m-d H:i:s'))->get();
+                if ($equals){
+                    foreach ($equals as $item){
+                        if ($item->id != $this->id){
+                            if ($this->hourlyMicrocontrollerData()->exists()) {
+                                $this->hourlyMicrocontrollerData()->forceDelete();
+                            }
+                            if ($this->dailyMicrocontrollerData()->exists()) {
+                                $this->dailyMicrocontrollerData()->forceDelete();
+                            }
+                            if ($this->clientAlert()->exists()) {
+                                $this->clientAlert()->forceDelete();
+                            }
+                            $this->forceDelete();
+                            return;
+                        }
+                    }
                 }
-                if ($this->dailyMicrocontrollerData()->exists()) {
-                    $this->dailyMicrocontrollerData()->forceDelete();
-                }
-                if ($this->clientAlert()->exists()) {
-                    $this->clientAlert()->forceDelete();
-                }
-                $this->forceDelete();
-                return;
+
             }
         }
         $this->saveQuietly();
